feature - database - deixando a criação opcional

+++ src/flask-arquitetura/src/infra/extensions/database_init.py
@@ -3,12 +3,17 @@ from src.domain.models.usuario import Usuario
 
 
 def init_app(app):
+    if not app.config.get('DATABASE_INIT'):
+        return
+
     with app.app_context():
         db.drop_all()
         db.create_all()
 
-        db.session.add(Usuario(nome='Fernando', email='fernandogjose@gmail.com', senha='12345', idade=41))
-        db.session.add(Usuario(nome='Priscila', email='priscilaecoaprender@gmail.com', senha='12345', idade=41))
+        db.session.add(Usuario(
+            nome='Fernando', email='fernandogjose@gmail.com', senha='12345', idade=41))
+        db.session.add(Usuario(
+            nome='Priscila', email='priscilaecoaprender@gmail.com', senha='12345', idade=41))
         db.session.commit()
 
         # users = db.session.execute(db.select(Usuario)).scalars()
